refactor(wrapped): extract story playback state machine into useStoryPlayback
Move stepIdx + year-reset, isPlaying, goNext/goBack/exit/restart,
auto-advance timer, keyboard nav, and touch-swipe verbatim into
_hooks/use-story-playback.ts (takes year + paused, owns its own
router). Exposes restart() instead of raw setStepIdx. Pruned now-unused
useCallback/useRef/STEP_DURATION_MS imports. Behavior-identical;
wrapped/page 496->438 LOC.
